Update Dungeon Maker App JavaScript
Implemented rooms, insertion of traps/treasure/enemies, and ability to select and change dungeon size. A button was added to randomly populate the generated dungeon with items (traps/treasure/enemies). This is separate from the button that generates a new dungeon.

Next step is to try to save the generated dungeon map in case the user generates one s/he likes.